+++ app/src/main/java/uz/texnopos/texnoposedufinance/data/model/Course.kt
@@ -4,7 +4,7 @@ data class Course(
     val id: String = "",
     val orgId: String = "",
     var name: String = "",
-    var price: Double = 0.0,
+    var price: Int = 0,
     var duration: Int = 0,
     var groups: List<Group> = listOf()
 )
